requirements.txt
requests==2.31.0
pandas==2.1.4
pandas-gbq==0.22.0
numpy==1.24.4
openpyxl==3.1.2
pyarrow==15.0.1
apache-airflow==2.8.3
apache-airflow-providers-apprise==1.2.2
apache-airflow-providers-google==10.16.0
airflow-provider-great-expectations==0.2.7
astronomer-cosmos==1.3.2
dbt-core==1.7.9
dbt-postgres==1.7.9
dbt-bigquery==1.7.6
soda-core-bigquery==3.3.5
streamlit==1.32.2
google-generativeai==0.4.1
python-dotenv==1.0.1
diagrams==0.23.4